pre: refactoring for deployment

+++ tcx_converter.py
@@ -2,6 +2,7 @@ from lxml import etree, objectify
 import requests
 import re
 import os
+import io
 import sys
 import csv
 import shutil
@@ -35,12 +36,22 @@ def parse_source(source, src_type="file"):
     elif src_type == "string":
         return etree.fromstring(source)
 
+def _fileOrBuffConstr(file_cont, isFile):
+    if (file_cont == 'data') and isFile:
+        return lambda: open(TEMP_DATA_FILE, 'w')
+    elif (file_cont == 'meta') and isFile:
+        return lambda: open(TEMP_META_FILE, 'w')
+    else:
+        return io.StringIO
+
 def convert_to_csv(source, src_type="file"):
+    isFile = src_type == 'file'
     #parse whole file
     try:
         root = parse_source(source, src_type)
     except Exception as e:
         print(e)
+        raise e
     # Activites node
     root = root.getchildren()[0]
     # write each activity in new file
@@ -52,9 +63,12 @@ def convert_to_csv(source, src_type="file"):
         datestring = parse_time(datestring).strftime('%Y%m%dT%H%M%SZ')
         try_filename = './' + datestring + '_' + act_name
         out_extension = '.csv'
+        if not isFile:
+            outL = list()
+            csvStr = ""
         # actual tabular data written to temporary file for merging with lap data
-        with open(TEMP_DATA_FILE, mode='w') as tmp_dataF:
-            csvwriter = csv.writer(tmp_dataF, lineterminator='\n')
+        with _fileOrBuffConstr('data', isFile)() as tmp_data:
+            csvwriter = csv.writer(tmp_data, lineterminator='\n')
             for i, lap in enumerate(activity.findall('{*}Lap')):
                 lap_n = str(i + 1)
                 metadata = []
@@ -106,10 +120,14 @@ def convert_to_csv(source, src_type="file"):
                         headerL = map(lambda s: headerDi.get(s, s), headerL)
                         csvwriter.writerow(headerL)
                     csvwriter.writerow(instL)
-            with open(TEMP_META_FILE, 'w') as metaF:
+            if not isFile:
+                csvStr += tmp_data.getvalue()
+            with _fileOrBuffConstr('meta', isFile)() as meta:
                 for s in metadata:
-                    metaF.write("# " + s + "\n")
-                metaF.write("#"*64 + "\n")
+                    meta.write("# " + s + "\n")
+                meta.write("#"*64 + "\n")
+                if not isFile:
+                    csvStr = meta.getvalue() + csvStr
             out_filename = try_filename + out_extension
             while os.path.exists(out_filename):
                 file_num_sep = '__'
@@ -121,12 +139,24 @@ def convert_to_csv(source, src_type="file"):
                     file_num_str = str(int(fileL[-1][:-len(out_extension)]) + 1)
                 out_filename = file_basename + file_num_sep + file_num_str + out_extension
             created_filenames.append(out_filename)
-        with open(out_filename, 'xb') as wfd:
-            for filename in TEMP_LIST:
-                with open(filename,'rb') as fd:
-                    shutil.copyfileobj(fd, wfd, 1024*1024*10)
-                os.remove(filename)
-        return created_filenames
+        if isFile:
+            with open(out_filename, 'xb') as wfd:
+                for filename in TEMP_LIST:
+                    with open(filename,'rb') as fd:
+                        shutil.copyfileobj(fd, wfd, 1024*1024*10)
+                    os.remove(filename)
+        else:
+            outL.append((out_filename, csvStr,))
+    if not isFile:
+        created_filenames = tuple((fname, csvStr) for fname, csvStr in outL)
+    return created_filenames
+
+
+
+def import_tcx(source, src_type):
+    csv_files = convert_to_csv(source, src_type)
+    dfTup = tuple(import_csv(fname, fname) for fname in csv_files)
+    return dfTup
 
 def main():
     in_filename = sys.argv[1]
